Rename AbstractCompactionStrategy to DelegatingCompactionStrategy, move to upper package
