+++ java/compaction-job-creation/src/main/java/sleeper/compaction/strategy/LeafPartitionCompactionStrategy.java
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.strategy.impl;
+package sleeper.compaction.strategy;
 
 import sleeper.compaction.job.CompactionJob;
+import sleeper.compaction.strategy.impl.CompactionFactory;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
 import sleeper.core.partition.Partition;
