+++ java/compaction-job-creation/src/main/java/sleeper/compaction/strategy/ShouldCreateJobsStrategy.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.strategy.impl;
+package sleeper.compaction.strategy;
 
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
