+++ java/compaction-job-creation/src/main/java/sleeper/compaction/strategy/impl/ShouldCreateJobsWithMaxPerPartition.java
@@ -17,6 +17,7 @@ package sleeper.compaction.strategy.impl;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import sleeper.compaction.strategy.ShouldCreateJobsStrategy;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
 import sleeper.configuration.properties.table.TableProperty;
